fix(check): probe the active embedding profile, not only the genesis seed
- tribal check now probes the active profile's endpoint (the live
  read/write identity, §5.7) when a corpus exists, falling back to the
  genesis seed only when there is no active profile, and the
  genesis-vs-active divergence note compares base_url and dimensions as
  well as provider and model (B2).
- The reindex check comment is reconciled from "stuck" to "a live
  reindex run", matching the shipped any-live-run warn condition (C8).
- The cumulative reindex_runs tallies (items/tags embedded and
  quarantined) widen to BIGINT to stay clear of the INT ceiling at scale;
  the single-snapshot *_enumerated columns stay INT (C6).
- Removed the pre-existing em dashes in scope.rs doc comments (C9).

+++ crates/tribal-domain/src/reindex.rs
@@ -103,12 +103,12 @@ pub struct ReindexRun {
     initiated_by_principal_id: PrincipalId,
     #[builder(default)]
     items_enumerated: Option<u32>,
-    items_embedded: u32,
+    items_embedded: i64,
     #[builder(default)]
     tags_enumerated: Option<u32>,
-    tags_embedded: u32,
-    items_quarantined: u32,
-    tags_quarantined: u32,
+    tags_embedded: i64,
+    items_quarantined: i64,
+    tags_quarantined: i64,
     #[builder(default)]
     error_message: Option<String>,
     #[builder(default)]
@@ -155,9 +155,9 @@ impl ReindexRun {
         self.items_enumerated
     }
 
-    /// Returns the count of item rows present for the profile.
+    /// Returns the cumulative count of item rows embedded across every pass.
     #[must_use]
-    pub fn items_embedded(&self) -> u32 {
+    pub fn items_embedded(&self) -> i64 {
         self.items_embedded
     }
 
@@ -167,21 +167,21 @@ impl ReindexRun {
         self.tags_enumerated
     }
 
-    /// Returns the count of tag rows present for the profile.
+    /// Returns the cumulative count of tag rows embedded across every pass.
     #[must_use]
-    pub fn tags_embedded(&self) -> u32 {
+    pub fn tags_embedded(&self) -> i64 {
         self.tags_embedded
     }
 
-    /// Returns the count of permanently-quarantined items.
+    /// Returns the cumulative count of permanently-quarantined items.
     #[must_use]
-    pub fn items_quarantined(&self) -> u32 {
+    pub fn items_quarantined(&self) -> i64 {
         self.items_quarantined
     }
 
-    /// Returns the count of permanently-quarantined tags.
+    /// Returns the cumulative count of permanently-quarantined tags.
     #[must_use]
-    pub fn tags_quarantined(&self) -> u32 {
+    pub fn tags_quarantined(&self) -> i64 {
         self.tags_quarantined
     }
